requirements.txt
requests
Flask
python-dotenv
urllib3==1.26.6
google-generativeai